requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
tiktoken
beautifulsoup4
Flask
gunicorn
python-dotenv